Create IssueCredentialsRepository
#82
