+++ data/src/main/java/by/alexandr7035/data/repository/IssueCredentialsRepositoryImpl.kt
@@ -1,0 +1,41 @@
+package by.alexandr7035.data.repository
+
+import by.alexandr7035.affinidi_id.domain.core.ErrorType
+import by.alexandr7035.affinidi_id.domain.model.credentials.issue_vc.IssueCredentialReqModel
+import by.alexandr7035.affinidi_id.domain.model.credentials.issue_vc.IssueCredentialResModel
+import by.alexandr7035.affinidi_id.domain.model.login.AuthStateModel
+import by.alexandr7035.affinidi_id.domain.repository.IssueCredentialsRepository
+import by.alexandr7035.data.core.AppError
+import by.alexandr7035.data.helpers.vc_issuance.VCIssuanceHelper
+import javax.inject.Inject
+
+class IssueCredentialsRepositoryImpl @Inject constructor(
+    private val vcIssuanceHelper: VCIssuanceHelper,
+) : IssueCredentialsRepository {
+
+    // 1) Build unsigned VC
+    // 2) Sign the VC
+    // 3) Store credential in the cloud wallet
+    // TODO give the user choice where to store in the future
+    override suspend fun issueCredential(
+        issueCredentialReqModel: IssueCredentialReqModel,
+        authState: AuthStateModel
+    ): IssueCredentialResModel {
+        try {
+            val unsignedVc = vcIssuanceHelper.buildUnsignedVC(issueCredentialReqModel)
+            val signedVc = vcIssuanceHelper.signCredential(unsignedVc, authState)
+            // Store only 1 VC, so just get last ID from response
+            val storedVCsID = vcIssuanceHelper.storeCredentials(listOf(signedVc), authState).last()
+
+            return IssueCredentialResModel.Success
+
+        } catch (appError: AppError) {
+            return IssueCredentialResModel.Fail(appError.errorType)
+        }
+        // Unknown exception
+        catch (e: Exception) {
+            e.printStackTrace()
+            return IssueCredentialResModel.Fail(ErrorType.UNKNOWN_ERROR)
+        }
+    }
+}
