+++ domain/src/main/java/by/alexandr7035/affinidi_id/domain/usecase/credentials/IssueCredentialUseCase.kt
@@ -4,12 +4,12 @@ import by.alexandr7035.affinidi_id.domain.core.ErrorType
 import by.alexandr7035.affinidi_id.domain.model.credentials.check_if_have_vc.CheckIfHaveVcReqModel
 import by.alexandr7035.affinidi_id.domain.model.credentials.issue_vc.IssueCredentialReqModel
 import by.alexandr7035.affinidi_id.domain.model.credentials.issue_vc.IssueCredentialResModel
-import by.alexandr7035.affinidi_id.domain.repository.CredentialsRepository
+import by.alexandr7035.affinidi_id.domain.repository.IssueCredentialsRepository
 import by.alexandr7035.affinidi_id.domain.usecase.user.GetAuthStateUseCase
 import javax.inject.Inject
 
 class IssueCredentialUseCase @Inject constructor(
-    private val credentialsRepository: CredentialsRepository,
+    private val issueCredentialsRepository: IssueCredentialsRepository,
     private val authStateUseCase: GetAuthStateUseCase,
     private val checkIfHaveCredentialUseCase: CheckIfHaveCredentialUseCase
 ) {
@@ -27,7 +27,7 @@ class IssueCredentialUseCase @Inject constructor(
             IssueCredentialResModel.Fail(ErrorType.ALREADY_HAVE_CREDENTIAL)
         } else {
             // Issue a new one
-            credentialsRepository.issueCredential(issueCredentialReqModel, authStateUseCase.execute())
+            issueCredentialsRepository.issueCredential(issueCredentialReqModel, authStateUseCase.execute())
         }
     }
 }
